backend/requirements.txt
# Core
fastapi[all]
# uvicorn[standard]
python-dotenv
python-multipart
edge-tts>=6.1.9

# PDF & Image Processing
pdf2image
opencv-python-headless
Pillow
numpy

# OCR
pytesseract

# AI/LLM
google-generativeai
google-cloud-pubsub
groq

# TTS
gtts
pydub

# Storage
supabase
httpx

# Utilities
requests
anyio
# AWS Lambda Adapter
mangum
awslambdaric
boto3